+++ elfo-core/src/lib.rs
@@ -23,6 +23,7 @@ pub use crate::{
 pub mod config;
 pub mod errors;
 pub mod messages;
+pub mod node;
 pub mod routers;
 pub mod signal;
 pub mod stream;
@@ -36,6 +37,7 @@ mod addr;
 mod address_book;
 mod context;
 mod demux;
+mod dumping;
 mod envelope;
 mod exec;
 mod group;
@@ -50,6 +52,21 @@ mod supervisor;
 
 #[doc(hidden)]
 pub mod _priv {
+    pub mod dumping {
+        pub use crate::dumping::*;
+
+        #[inline]
+        pub fn of<C, K, S>(context: &crate::Context<C, K, S>) -> &Dumper {
+            context.dumper()
+        }
+    }
+
+    pub mod node {
+        pub fn set_node_no(node_no: crate::node::NodeNo) {
+            crate::node::set_node_no(node_no)
+        }
+    }
+
     pub use crate::{
         envelope::{AnyMessageBorrowed, AnyMessageOwned, EnvelopeBorrowed, EnvelopeOwned},
         message::{AnyMessage, LocalTypeId, MessageVTable, MESSAGE_LIST},